Isolate scraper access test configuration
Move pure scraper options out of runtime orchestration so connectivity tests do not load notifier secrets, and explicitly use permissive application filtering for the public-site access probe.


Copilot-Session: b4455f2e-c505-4ea0-8a73-cbc3fb022a7f
