+++ apps/scraper/src/scraperOptions.ts
@@ -1,0 +1,12 @@
+import type { ScraperOptions } from "israeli-bank-scrapers";
+
+export const scraperOptions: Partial<ScraperOptions> = {
+  navigationRetryCount: 3,
+  viewportSize: { width: 1920, height: 1080 },
+  optInFeatures: [
+    "mizrahi:pendingIfHasGenericDescription",
+    "mizrahi:pendingIfNoIdentifier",
+    "mizrahi:pendingIfTodayTransaction",
+    "isracard-amex:skipAdditionalTransactionInformation",
+  ],
+};
